refactor: organize LiFi-specific components under _ui/lifi/
All cross-chain-swap files that are LiFi-specific move into _ui/lifi/
and drop the redundant `cross-chain-swap-` filename prefix (the folder
already communicates scope, mirroring the existing _ui/near-intents/
pattern).

Includes the LiFi xswap provider, trade-review-dialog and its hooks,
all token/trade widgets, the chain-unsupported message, the LiFi
maintenance hook + message, and the LiFi-side switch-tokens button.

Shared at root: cross-chain-swap-widget (orchestrator),
cross-chain-swap-settings-overlay, xswap-form-provider,
derivedstate-swap-helpers (parent _ui/).

Pure rename + import-path refactor; no behavior or component-name
changes.
