+++ apps/web/src/app/(networks)/(evm)/[chainId]/(trade)/cross-chain-swap/_ui/lifi/trade-review-dialog/confirm-swap-button.tsx
@@ -7,7 +7,7 @@ import {
 } from '@sushiswap/telemetry'
 import { Button, Dots } from '@sushiswap/ui'
 import type { LifiXSwapSupportedChainId } from 'src/config'
-import { useLifiXSwap } from '../lifi-xswap-provider'
+import { useLifiXSwap } from '../xswap-provider'
 
 export function ConfirmSwapButton<
   TChainId0 extends LifiXSwapSupportedChainId,
