+++ apps/web/src/app/(networks)/(evm)/[chainId]/(trade)/cross-chain-swap/_ui/lifi/trade-review-dialog/recipient-section.tsx
@@ -2,7 +2,7 @@
 
 import { List } from '@sushiswap/ui'
 import { getChainById, shortenAddress } from 'sushi'
-import { useLifiXSwap } from '../lifi-xswap-provider'
+import { useLifiXSwap } from '../xswap-provider'
 
 export function RecipientSection() {
   const {
