+++ apps/web/src/app/(networks)/(evm)/[chainId]/(trade)/cross-chain-swap/_ui/lifi/trade-review-dialog/review-intro.tsx
@@ -6,7 +6,7 @@ import type { LifiXSwapSupportedChainId } from 'src/config'
 import { nativeFromChainId } from 'src/lib/currency-from-chain-id'
 import { getChainById } from 'sushi'
 import { stringify } from 'viem'
-import { useLifiXSwap } from '../lifi-xswap-provider'
+import { useLifiXSwap } from '../xswap-provider'
 
 export function ReviewIntro<
   TChainId0 extends LifiXSwapSupportedChainId,
