+++ apps/web/src/app/(networks)/(evm)/[chainId]/(trade)/cross-chain-swap/_ui/lifi/xswap-provider.tsx
@@ -25,7 +25,7 @@ import {
   type XSwapFormMutators,
   type XSwapFormStateValues,
   useXSwapForm,
-} from './xswap-form-provider'
+} from '../xswap-form-provider'
 
 type NewTokenInput = Parameters<typeof newToken>[0]
